added si1145 init operation choice cleanup sensor api data types

+++ IoTSense/Inc/th02.h
@@ -8,8 +8,7 @@
 #ifndef TH02_H_
 #define TH02_H_
 
-uint32_t th02_get_humidity();
-
+uint16_t th02_get_humidity();
 uint16_t th02_get_temperature();
 
 #endif /* TH02_H_ */
